% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classify_splits}
\alias{classify_splits}
\title{\code{classify_splits} classifies samples using centroids}
\usage{
classify_splits(d2, L, opt = NULL)
}
\arguments{
\item{d2}{A data.frame containing gene (rows) expression from the samples (columns)
to be classified;}

\item{L}{A list containing centroids produced by \code{computeCentroids}
function;}

\item{opt}{Determine which split to be computed: ss: sensory / secretory split;
la : Luminal / Apocrine split;}
}
\value{
\code{classify_splits} Returns a data.frame with columns
\item{scores}{Distance to centroids}
\code{"sensory","secretory"} if \code{opt} set to ss or
\code{"Luminal","MA"} if \code{opt} set to la
\item{predictions}{Centroid association}
\code{"pred.ss","pred.ss.strict"} if \code{opt} set to ss : sensory / Basal
\code{"pred.la","pred.la.strict"} if \code{opt} set to la : Luminal,
MA: Molecular Apocrine
The strict version annotates the samples too far from any
centroid as unclassified
}
\description{
\code{classify_splits} classifies samples using centroids
}
