% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LABClassifier.R
\name{compute.cutoff.distrib}
\alias{compute.cutoff.distrib}
\title{\code{compute.cutoff.distrib} uses \code{densityMclust} from mclust package.}
\usage{
compute.cutoff.distrib(data, prefix, feature, plot = F)
}
\arguments{
\item{data}{Either a data.frame storing in columns the data to evaluate or a
numeric vector}

\item{prefix}{The prefix for the plot file that will be stored in the Output
folder}

\item{feature}{If data is a data.frame, feature is the column name of interest.
In any case feature will be added to the plot file name}

\item{plot}{If TRUE, will plot the distribution of the data and the predicted cutoff}
}
\value{
\code{compute.cutoff.distrib} Returns a numerical cutoff.
}
\description{
\code{compute.cutoff.distrib} uses \code{densityMclust} from mclust package.
}
