% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LABClassifier.R
\name{import.data.zenodo}
\alias{import.data.zenodo}
\title{\code{import.data.zenodo} dowloads files from zenodo (doi:10.5281/zenodo.10935179)
and returns a list of data.frame containing expression
values from cohorts that can be used to test the classifier}
\usage{
import.data.zenodo(dataset = "all")
}
\arguments{
\item{dataset}{A character vector containing one or more dataset identifiers
from "TCGA", "METABRIC", "ICGC", "EORTC" or "all" if all four datasets are wanted.
The zenodo files are stored in "./ext_data" once downloaded.}
}
\value{
A list of data.frames containing gene expression
}
\description{
\code{import.data.zenodo} dowloads files from zenodo (doi:10.5281/zenodo.10935179)
and returns a list of data.frame containing expression
values from cohorts that can be used to test the classifier
}
