% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centroides.R
\name{lab.classifier}
\alias{lab.classifier}
\title{\code{lab.classifier} computes splits according to either average expression
of TFs defining the splits or their enrichment according to ssGSEA}
\usage{
lab.classifier(
  data,
  dir.path = ".",
  prefix,
  method = "mean",
  sensor.genes = NULL,
  secretor.genes = NULL,
  asc.genes = NULL,
  lsc.genes = NULL,
  use.ra = F
)
}
\arguments{
\item{data}{A data.frame containing expression values from samples to classify;}

\item{dir.path}{A character string defining the path to where the results
will be stored. Default: current directory.}

\item{prefix}{A character string: prefix of the outputs.}

\item{method}{A character string: Either \code{mean} or \code{ssgsea}. \code{mean} computes the average
expression values of TFs defining the splits and \code{ssgsea} uses the function gsva
from GSVA R packages with no normalisation. Default: mean}

\item{sensor.genes, secretor.genes, asc.genes, lsc.genes}{Vectors: Gene list to compute splits.
If no gene list is given, pre-defined ones are used. Default: NULL}

\item{use.ra}{Boolean: If set to TRUE, the Androgen Receptor target genes are added to the molecular
apocrine split genes. Default: FALSE}
}
\value{
A data.frame containing LAB classification
}
\description{
\code{lab.classifier} computes splits according to either average expression
of TFs defining the splits or their enrichment according to ssGSEA
}
