% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LABclassifier}
\alias{LABclassifier}
\title{\code{LABclassifier} helps in visualising the sample classification}
\usage{
LABclassifier(data, LABClassif = NULL, raw.counts = F, plot = F)
}
\arguments{
\item{data}{A data.frame containing gene (rows) expression from the samples (columns)
to be classified}

\item{LABClassif}{An object of class \code{LABclassifier} obtained with
\code{make.LABclassifier} function. If NULL, the precomputed classifier
is used (default);}

\item{raw.counts}{A logical. Set it to TRUE if you are using raw sequencing
data, then they will be transformed in TPM. Default: FALSE;}

\item{plot}{A logical. Set it to TRUE to display classification
diagnostic plots. Default: FALSE}
}
\value{
\code{LABclassifier} Returns a data.frame with columns
\item{sensory2secretory}{Distance to sensory centroid}
\item{Luminal2Apocrine}{Distance to Luminal centroid}
\item{prediction}{Centroid association: Luminal, Basal, MA: Molecular Apocrine}
\item{prediction.strict}{Centroid association with annotation of the samples
too far from any centroid as unclassified}
}
\description{
\code{LABclassifier} helps in visualising the sample classification
}
\examples{
data(TCGA.rsem)
LABclassifier(TCGA.rsem,plot=TRUE)
}
