% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centroides.R
\name{make.LABclassifier}
\alias{make.LABclassifier}
\title{\code{make.LABclassifier} compute centroids}
\usage{
make.LABclassifier(data, annot)
}
\arguments{
\item{data}{A data.frame containing expression values from classified samples;}

\item{annot}{A data.frame containing the Luminal/Basal/MA annotations
for samples}
}
\value{
A object of class \code{LABclassifier} containing a list
\item{centroid.ss}{Centroids to classify according to the sensory/secretor split}
\item{centroid.la}{Centroids to classify according to the Luminal/Apocrine split}
}
\description{
\code{make.LABclassifier} compute centroids
}
