% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50.robust}
\alias{pam50.robust}
\title{Object from genefu package to compute PAM50 classification}
\format{
\subsection{\code{pam50.robust}}{

#' \describe{
See genefu package (Deena M.A. Gendoo et al. (2021). genefu: Computation of Gene
Expression-Based Signatures in Breast Cancer. R package version 2.26.0.)
}
}
}
\source{
\url{http://www.pmgenomics.ca/bhklab/software/genefu}
}
\usage{
pam50.robust
}
\description{
Object from genefu package to compute PAM50 classification
}
\keyword{datasets}
