% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PAMgenefu}
\alias{PAMgenefu}
\title{\code{PAMgenefu} uses \code{molecular.subtyping} from genefu package.
Please cite the corresponding paper (See reference).}
\usage{
PAMgenefu(data)
}
\arguments{
\item{data}{A data.frame containing gene (gene symbols as rows) expression
from the samples (sample IDs as columns) to be classified or a  vector
containing gene expression named with gene SYMBOL}
}
\value{
\code{PAMgenefu} Returns a data.frame with 1 column
\item{PAM50}{PAM50 classification: LumA and LumB (Luminal A and B), Basal, Her2.}
}
\description{
\code{PAMgenefu} uses \code{molecular.subtyping} from genefu package.
Please cite the corresponding paper (See reference).
}
\references{
genefu:  Deena M.A. Gendoo et al. (2021). genefu: Computation of Gene
Expression-Based Signatures in Breast Cancer. R package version 2.26.0.
http://www.pmgenomics.ca/bhklab/software/genefu
}
