% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LABClassifier.R
\name{plot_splits}
\alias{plot_splits}
\title{\code{plot_splits} helps visualise splits}
\usage{
plot_splits(filename, LP.dat, LA.dat)
}
\arguments{
\item{filename}{set the filename prefix for the plot}

\item{LP.dat, LA.dat}{are results from the ssGSEA scoring}
}
\description{
\code{plot_splits} helps visualise splits
}
