% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.rsem}
\alias{TCGA.rsem}
\title{Normalized expression data from TCGA BRCA samples}
\format{
\subsection{\code{TCGA.rsem}}{

A data frame with 48 genes (rows) and 674 samples (columns):
\describe{
\item{genes}{Genes useful to classify samples with centroids.}
\item{samples}{Breast cancer samples from the TCGA consortium filtered
for tumor stage II and III}
\item{expression}{The values are RSEM normalized}
}
}
}
\source{
\url{https://tcga-xena-hub.s3.us-east-1.amazonaws.com/download/TCGA.BRCA.sampleMap\%2FHiSeqV2.gz}
}
\usage{
TCGA.rsem
}
\description{
Normalized expression data from TCGA BRCA samples
}
\keyword{datasets}
