% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LABclassifier}
\alias{LABclassifier}
\title{\code{LABclassifier} helps in visualising the sample classification}
\usage{
LABclassifier(
  data,
  LABClassif = NULL,
  raw.counts = F,
  log2T = F,
  id.type = "SYMBOL",
  PAM50 = F,
  plot = F,
  prefix = NULL
)
}
\arguments{
\item{data}{A data.frame containing gene (gene symbols as rows) expression
from the samples (sample IDs as columns) to be classified or a  vector
containing gene expression named with gene SYMBOL}

\item{LABClassif}{An object of class \code{LABclassifier} obtained with
\code{make.LABclassifier} function. If NULL, the precomputed classifier
is used (default);}

\item{raw.counts}{A logical. Set it to TRUE if you are using raw sequencing
data, then they will be transformed in TPM. Default: FALSE;}

\item{log2T}{A logical. If true, data are log2 transformed. Default: FALSE;}

\item{id.type}{A character string. Gene identifier type in the input matrix. Possible values are
ensEMBL, entrezID, SYMBOL (default)}

\item{PAM50}{A logical. Compute PAM50 classification using \code{molecular.subtyping}
function from genefu package. If used, please cite the corresponding paper (See reference).}

\item{plot}{A logical. Set it to TRUE to display classification
diagnostic plots. Default: FALSE}

\item{prefix}{A character string. If set, a directory Output is created in the
working directory (if not exists), and the prediction result is saved in as a
tabulated file. The filename is prefix_LAB_prefix.tab}
}
\value{
\code{LABclassifier} Returns a data.frame with columns
\item{sensory2secretory}{Distance to sensory centroid}
\item{Luminal2Apocrine}{Distance to Luminal centroid}
\item{prediction}{Centroid association: Luminal, Basal, MA: Molecular Apocrine}
\item{prediction.strict}{Centroid association with annotation of the samples
too far from any centroid as unclassified}
\item{confidence}{Summarizes the distance to centroids, it is just an indicator:
\itemize{
\item{high confidence}{ : 0.75 < scores > 1.15}
\item{medium confidence}{ : 0.75 > scores < 1.25}
\item{low confidence{ : 0.9 > scores < 1.1 (unclassified samples in strict predictions)}
}}}
}
\description{
\code{LABclassifier} helps in visualising the sample classification
}
\examples{
data(TCGA.rsem)
LABclassifier(TCGA.rsem,plot=TRUE)
}
\references{
genefu:  Deena M.A. Gendoo et al. (2021). genefu: Computation of Gene
Expression-Based Signatures in Breast Cancer. R package version 2.26.0.
http://www.pmgenomics.ca/bhklab/software/genefu
}
