% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LABClassifier.R
\name{LABclassifier}
\alias{LABclassifier}
\title{\code{LABclassifier} computes splits according expression
of TFs defining the splits enrichment according to ssGSEA
\code{LABclassifier} helps in visualising the sample classification}
\usage{
LABclassifier(
  data,
  dir.path = ".",
  prefix = "myClassif",
  raw.counts = F,
  log2T = F,
  id.type = "SYMBOL",
  PAM50 = F,
  plot = T,
  sensor.genes = NULL,
  secretor.genes = NULL,
  asc.genes = NULL,
  lsc.genes = NULL,
  genes.for.heatmap = NULL,
  colorBlind = F
)
}
\arguments{
\item{data}{A data.frame containing gene (gene symbols as rows) expression
from the samples (sample IDs as columns) to be classified or a  vector
containing gene expression named with gene SYMBOL}

\item{dir.path}{A character string defining the path to where the results
will be stored. Default: current directory.}

\item{prefix}{A character string: prefix of the outputs. Default: myClassif}

\item{raw.counts}{A logical. Set it to TRUE if you are using raw sequencing
data, then they will be transformed in TPM. Default: FALSE;}

\item{log2T}{A logical. If true, data are log2 transformed. Default: FALSE;}

\item{id.type}{A character string. Gene identifier type in the input matrix. Possible values are
ensEMBL, entrezID, SYMBOL (default)}

\item{PAM50}{A logical. Compute PAM50 classification using \code{molecular.subtyping}
function from genefu package. If used, please cite the corresponding paper (See reference).}

\item{plot}{A logical. Set it to TRUE to display classification
diagnostic plots. Default: FALSE}

\item{sensor.genes, secretor.genes, asc.genes, lsc.genes}{Vectors: Gene list to compute splits.
If no gene list is given, pre-defined ones are used. Default: NULL}

\item{genes.for.heatmap}{A gene symbol vector to be plotted in the heatmap. If NULL,
genes used for splitting estimation are used. Default=NULL}

\item{colorBlind}{If set to TRUE, a color palette color blind friendly is used. Default: FALSE}
}
\value{
A data.frame containing LAB classification
}
\description{
\code{LABclassifier} computes splits according expression
of TFs defining the splits enrichment according to ssGSEA
\code{LABclassifier} helps in visualising the sample classification
}
\examples{
\dontrun{
# dataset is a character vector containing one or more dataset identifiers
from "TCGA", "METABRIC", "ICGC", "EORTC" or "all" if all four datasets are wanted.
data <- import.data.zenodo(dataset = "TCGA")
# If the data were already dowloaded, they are stored in "./ext_data" and can
# be accessed using read.table("./ext_data/TCGA_matrix.txt",header=T,row.names=1)
# Test on TCGA data
LABclassifier(data$TCGA,plot=TRUE)
}
}
\references{
genefu:  Deena M.A. Gendoo et al. (2021). genefu: Computation of Gene
Expression-Based Signatures in Breast Cancer. R package version 2.26.0.
http://www.pmgenomics.ca/bhklab/software/genefu
}
