% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LABClassifier.R
\docType{data}
\name{pam50.robust}
\alias{pam50.robust}
\title{PAM50 Robust Dataset}
\format{
A list containing robust centroids and associated metadata.
}
\usage{
data(pam50.robust)
}
\description{
This dataset contains the robust centroids used for PAM50 molecular subtyping
from the genefu package.
}
\examples{
data(pam50.robust)
}
\keyword{datasets}
